requirements.txt
Flask
mysql-connector-python
python-dotenv
requests
transformers
sqlparse
FLASK_CORS
Flask_session
python-dateutil
groq